convert MultipleClients test to be parametrized
